moved poolShare helpers to the app level
